# contracts/openapi/mediaflow.v1.yaml
openapi: 3.1.2
info:
  title: MediaFlow API
  version: v1
  description: Contract-first API for MediaFlow identity, inbox scanning, continuous discovery, deterministic identification, review administration, formal catalog reads, downloader management, and source automation.
servers:
  - url: /
paths:
...
  /api/v1/deployment-roots:
    get:
      operationId: listDeploymentRoots
      summary: List deployment capability roots
      description: Returns operator-configured roots that bound which local paths may be selected by inbox directories.
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/DeploymentRootList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/inbox-directories/preflight:
    post:
      operationId: preflightInboxDirectory
      summary: Preflight an inbox directory
      description: Normalizes and checks a capability-root-relative path for readability and overlap without persisting an inbox directory.
      security: *write-security
      parameters: *write-parameters
      requestBody: { $ref: '#/components/requestBodies/InboxDirectoryPreflightRequest' }
      responses:
        '200': { $ref: '#/components/responses/InboxDirectoryPreflight' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/inbox-directories:
    get:
      operationId: listInboxDirectories
      summary: List inbox directories
      description: Returns a bounded cursor page of persisted inbox directories and their last observed health.
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200': { $ref: '#/components/responses/InboxDirectoryPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createInboxDirectory
      summary: Create an inbox directory
      description: Persists a validated capability-root-relative path as a scan source after authoritative containment and overlap checks.
      security: *write-security
      parameters: *write-parameters
      requestBody: { $ref: '#/components/requestBodies/CreateInboxDirectoryRequest' }
      responses:
        '201': { $ref: '#/components/responses/InboxDirectory' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/inbox-directories/{inboxDirectoryId}:
    parameters: [{ $ref: '#/components/parameters/InboxDirectoryId' }]
    get:
      operationId: getInboxDirectory
      summary: Read an inbox directory
      description: Returns one persisted inbox directory and its last observed availability projection.
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/InboxDirectory' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/v1/inbox-directories/{inboxDirectoryId}/discovery-policy:
    parameters: [{ $ref: '#/components/parameters/InboxDirectoryId' }]
    get:
      operationId: getDiscoveryPolicy
      summary: Read an inbox discovery policy
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/DiscoveryPolicy' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      operationId: putDiscoveryPolicy
      summary: Replace an inbox discovery policy
      security: *write-security
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/Origin' }
        - { $ref: '#/components/parameters/SecFetchSite' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody: { $ref: '#/components/requestBodies/PutDiscoveryPolicyRequest' }
      responses:
        '200': { $ref: '#/components/responses/DiscoveryPolicy' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: __Host-mediaflow_session }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-Token }
  parameters:
    Cursor: { name: cursor, in: query, schema: { type: string, minLength: 1, maxLength: 512 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
    InboxDirectoryId: { name: inboxDirectoryId, in: path, required: true, schema: { type: string, format: uuid } }
...
    CsrfToken: { name: X-CSRF-Token, in: header, required: true, schema: { type: string, minLength: 1 } }
    Origin: { name: Origin, in: header, required: false, description: Exact public origin proof used when Fetch Metadata is absent and checked whenever supplied., schema: { type: string, format: uri } }
    SecFetchSite: { name: Sec-Fetch-Site, in: header, required: false, description: Browser Fetch Metadata primary same-origin proof; cross-site and malformed values fail closed., schema: { type: string, enum: [same-origin, same-site, cross-site, none] } }
...
    IfMatch: { name: If-Match, in: header, required: true, description: Decimal configuration version expected by the caller., schema: { type: string, pattern: '^[0-9]{1,19}$' } }
...
  requestBodies:
...
    InboxDirectoryPreflightRequest: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InboxDirectoryPreflightRequest' } } } }
    CreateInboxDirectoryRequest: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateInboxDirectoryRequest' } } } }
...
    PutDiscoveryPolicyRequest: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PutDiscoveryPolicyRequest' } } } }
...
  responses:
...
    DeploymentRootList: { description: Deployment roots, content: { application/json: { schema: { $ref: '#/components/schemas/DeploymentRootList' } } } }
    InboxDirectoryPreflight: { description: Inbox preflight, content: { application/json: { schema: { $ref: '#/components/schemas/InboxDirectoryPreflight' } } } }
    InboxDirectoryPage: { description: Inbox directories, content: { application/json: { schema: { $ref: '#/components/schemas/InboxDirectoryPage' } } } }
    InboxDirectory: { description: Inbox directory, content: { application/json: { schema: { $ref: '#/components/schemas/InboxDirectory' } } } }
...
    DiscoveryPolicy: { description: Versioned inbox discovery policy, content: { application/json: { schema: { $ref: '#/components/schemas/DiscoveryPolicy' } } } }
...
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    Forbidden: { description: CSRF or Origin rejected, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    Conflict: { description: Request conflicts with current state, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    ValidationFailed: { description: Request validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
...
    ServiceUnavailable: { description: Service is not ready, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    InternalError: { description: Internal error, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
  schemas:
    ApiErrorBody:
      description: Allow-listed API failure envelope with a correlation request ID and optional safe structured details.
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code: { type: string, enum: [bootstrap.already_completed, bootstrap.invalid_secret, auth.invalid_credentials, auth.rate_limited, session.expired, csrf.invalid, origin.untrusted, root.not_found, root.unavailable, path.invalid, path.escape, path.symlink_forbidden, inbox.overlap, inbox.not_found, task.not_found, task.invalid_state, task.lease_lost, integration.not_configured, integration.unauthorized, integration.rate_limited, integration.unsupported-version, provider.unavailable, provider.timeout, provider.response-too-large, provider.invalid-response, download.correlation-ambiguous, download.remote-missing, automation.source-disabled, automation.signature-invalid, automation.replay, automation.action-invalid, automation.payload-invalid, automation.downstream-conflict, organization.root-read-only, organization.target-overlap, organization.target-unavailable, organization.path-outside-root, organization.target-exists, organization.source-changed, organization.hardlink-cross-device, organization.operation-unsupported, organization.journal-ambiguous, revision.changed, event.cursor_expired, request.conflict, validation.failed, internal.error] }
            message: { type: string }
            request_id: { type: string, format: uuid }
            details: { type: object, additionalProperties: { type: [string, number, boolean, 'null'] } }
...
    DeploymentRootList:
      description: Bounded collection of operator-configured filesystem capability roots.
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties: { items: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/DeploymentRoot' } }, next_cursor: { type: [string, 'null'] } }
    DeploymentRoot:
      description: Opaque capability-root identity, display label, and permitted access mode without a host absolute path.
      type: object
      additionalProperties: false
      required: [id, label, access]
      properties: { id: { type: string }, label: { type: string }, access: { enum: [read-only, read-write] } }
    InboxDirectoryPreflightRequest:
      description: Capability-root identity and root-relative path to check without persisting state.
      type: object
      additionalProperties: false
      required: [root_id, relative_path]
      properties: { root_id: { type: string }, relative_path: { type: string, minLength: 1 } }
    InboxDirectoryPreflight:
      description: Normalized root-relative path together with authoritative readability and overlap observations.
      type: object
      additionalProperties: false
      required: [root_id, relative_path, readable, overlaps_existing]
      properties: { root_id: { type: string }, relative_path: { type: string }, readable: { type: boolean }, overlaps_existing: { type: boolean } }
    CreateInboxDirectoryRequest:
      description: Capability-root identity and root-relative path submitted for authoritative validation and persistence.
      type: object
      additionalProperties: false
      required: [root_id, relative_path]
      properties: { root_id: { type: string }, relative_path: { type: string, minLength: 1 } }
    InboxDirectory:
      description: Persisted scan-source identity, normalized root-relative path, and last observed availability.
      type: object
      additionalProperties: false
      required: [id, root_id, relative_path, health, last_checked_at]
      properties: { id: { type: string, format: uuid }, root_id: { type: string }, relative_path: { type: string }, health: { enum: [available, unavailable] }, last_checked_at: { type: string, format: date-time } }
    InboxDirectoryPage:
      description: Bounded cursor page of persisted inbox-directory projections.
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties: { items: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/InboxDirectory' } }, next_cursor: { type: [string, 'null'] } }
...
    DiscoveryPolicy:
      description: Versioned stability and reconciliation policy for one inbox directory.
      type: object
      additionalProperties: false
      required: [inbox_directory_id, minimum_age_seconds, stable_observation_interval_seconds, reconcile_interval_seconds, watcher_enabled, config_version]
      properties:
        inbox_directory_id: { type: string, format: uuid }
        minimum_age_seconds: { type: integer, minimum: 0, maximum: 86400, default: 60 }
        stable_observation_interval_seconds: { type: integer, minimum: 1, maximum: 86400, default: 30 }
        reconcile_interval_seconds: { type: integer, minimum: 60, maximum: 604800, default: 900 }
        watcher_enabled: { type: boolean, default: true }
        config_version: { type: integer, minimum: 1, maximum: 9007199254740991 }
    PutDiscoveryPolicyRequest:
      description: Replacement stability and reconciliation settings for one inbox directory.
      type: object
      additionalProperties: false
      required: [minimum_age_seconds, stable_observation_interval_seconds, reconcile_interval_seconds, watcher_enabled]
      properties:
        minimum_age_seconds: { type: integer, minimum: 0, maximum: 86400, default: 60 }
        stable_observation_interval_seconds: { type: integer, minimum: 1, maximum: 86400, default: 30 }
        reconcile_interval_seconds: { type: integer, minimum: 60, maximum: 604800, default: 900 }
        watcher_enabled: { type: boolean, default: true }
...